refactor: split screens.tsx into per-concern section files
screens.tsx (1971 lines) violated AGENTS.md §3 (~500 line limit).
Split by concern — each section is an independent component with no
behavioral change. screen-helpers.tsx holds shared primitives
(BarMetric, CoffeeStepper, EmptyState, format/band helpers) used by
multiple sections. screens.tsx is now a thin re-export barrel so all
existing imports in App.tsx continue to work unchanged.

Addresses #136
